Inline job_queue_node_status_transition
